Refactor comm. and compression backend in 1-bit adam.
